% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efronData.R
\docType{data}
\name{efron}
\alias{efron}
\title{The standard set of Efron's dice}
\format{A data table with 4 columns. Each column represents a die
with six faces.}
\usage{
data(efron)
}
\description{
A dataset containing the four standard Efron's dice, non-transitive
set of dice with winning probabilities of 2/3.
}
\examples{
data(efron)
is.nonTransitive(efron, prob = 2/3)
}
\keyword{datasets}

