% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miwinData.R
\docType{data}
\name{miwin}
\alias{miwin}
\title{Miwin set of dice.}
\format{A data table with 3 columns. Each column represents a die with six faces.}
\usage{
data(miwin)
}
\description{
A dataset containing the Miwin set of dice. This set is non-transitively losing, with losing probabilities of 17/36.
}
\examples{
data(miwin)
is.nonTransitive(miwin, prob = 17/36)
}
\keyword{datasets}

