% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oskarData.R
\docType{data}
\name{oskar}
\alias{oskar}
\title{Oskar van Deventer set of dice.}
\format{A data table with 6 columns. Each column represents a die
with six faces.}
\usage{
data(oskar)
}
\description{
A dataset containing the Oskar van Deventer dice, non-transitive
set of dice where  A beats {B,C,E}; B beats {C,D,F}; C beats {D,E,G}; #' D beats {A,E,F}; E beats {B,F,G}; F beats {A,C,G}; G beats {A,B,D}. #' Consequently, for arbitrarily chosen two dice there is a third one
that beats both of them.
}
\examples{
data(oskar)
is.nonTransitive(oskar)
}
\keyword{datasets}

